Migrate account value from int to double
- Now any monetary values are represented by Double instead of Integer.

+++ HDSLedger/Client/src/main/java/pt/ulisboa/tecnico/hdsledger/client/Client.java
@@ -73,7 +73,7 @@ public class Client {
                             System.out.print(">> ");
                             break;
                         }
-                        clientLibrary.transfer(id, splitInput[1], Integer.valueOf(splitInput[2]));
+                        clientLibrary.transfer(id, splitInput[1], Double.valueOf(splitInput[2]));
                         break;
                     case "balance": case "b":
                         clientLibrary.check_balance();
